requirements.txt
Django==1.5.1
South==0.8.1
argparse==1.2.1
beautifulsoup4==4.2.1
dateutils==0.6.6
distribute==0.6.34
django-appconf==0.6
django-bootstrap-toolkit==2.12.0
django-compressor==1.3
django-qrcode==0.3
feedparser==5.1.3
lxml==3.2.1
psycopg2==2.5
python-dateutil==2.1
pytz==2013b
qrencode==1.01
simplejson==3.3.0
six==1.3.0
stevedore==0.8
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
wsgiref==0.1.2
yolk==0.4.3
